refactor out some logic into buildSrc
