+++ buildSrc/build.gradle.kts
@@ -4,7 +4,10 @@ plugins {
     id("com.diffplug.spotless").version("6.21.0")
 }
 
-dependencies { implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3") }
+dependencies {
+    implementation("org.cyclonedx:cyclonedx-core-java:9.0.5")
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3")
+}
 
 spotless {
     kotlinGradle {
